Moved Sockets to Standalone File
- Moved function definitions for server-side sockets connection to standalone file
- Implemented import in `app.py` to reflect new file addition
